requirements.txt
streamlit>=1.28.1
openai>=1.3.6
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-dotenv>=1.0.0
numpy>=1.24.4